[Fix] field quote now is in long [Add] CatCode Deserializer [Add] Ping pong command

+++ src/main/java/cc/sfclub/mirai/packets/GroupQuoteMessage.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Builder
 public class GroupQuoteMessage extends Packet {
-    private int quote;
+    private long quote;
     private String sessionKey;
     private long target;
     private List<MiraiTypeMessage> messageChain;
